LUS-9560 tests: add ior_CLEANUP
Sometimes we need to run read IOR test separately from
write test. IOR -r test requires file exists before test starts.
If ior_CLEANUP set to false $testdir is not removed and
$testdir/iorData file can be used for IOR -r further test.

Test-Parameters: testlist=parallel-scale/ior.*
Reviewed-on: https://es-gerrit.dev.cray.com/158091
Tested-by: Jenkins Build User <nssreleng@cray.com>

+++ lustre/tests/functions.sh
@@ -609,6 +609,7 @@ run_ior() {
 	ior_xferSize=${ior_xferSize:-1M}
 	ior_type=${ior_type:-POSIX}
 	ior_DURATION=${ior_DURATION:-30}        # minutes
+	ior_CLEANUP=${ior_CLEANUP:-true}
 	local multiplier=1
 	case ${ior_blockUnit} in
 		[G])
@@ -687,7 +688,7 @@ run_ior() {
     if [ $rc != 0 ] ; then
         error "ior failed! $rc"
     fi
-    rm -rf $testdir
+    $ior_CLEANUP && rm -rf $testdir || true
 }
 
 run_mib() {
